requirements.txt
mss
Pillow
pynput
pywin32
pyinstaller
